+++ R/utils.R
@@ -13,9 +13,17 @@ warning2 = function(...) warning(..., call. = FALSE)
 #' @examples
 #' z = structure(list(a = 1), foo = 2)
 #' base::attr(z, 'f')  # 2
-#' xfun::attr(z, 'f')  # NULL
-#' xfun::attr(z, 'foo')  # 2
-attr = function(...) base::attr(..., exact = TRUE)
+#' xfun::attr2(z, 'f')  # NULL
+#' xfun::attr2(z, 'foo')  # 2
+attr2 = function(...) base::attr(..., exact = TRUE)
+
+#' @rdname attr2
+#' @export
+#' @note `xfun::attr()` is deprecated. Please use `xfun::attr2()` instead.
+attr = function(...) {
+  .Deprecated('xfun::attr2()', old = 'xfun::attr()')
+  attr2(...)
+}
 
 #' Set environment variables
 #'